feat(client): add full flow for adding and redirecting urls
Wires the client-side form to the Go API endpoints. Adds a dynamic route
with a loading screen to redirect shortened URLs.

+++ services/rest/internal/handler/url_mapping.go
@@ -102,7 +102,7 @@ func GetURLMappingHandler(client *redis.Client, validate *validator.Validate) ht
 		w.Header().Set("Content-Type", "application/json")
 		json.NewEncoder(w).Encode(map[string]string{
 			"message": "Retrieved actual URL",
-			"short":   val,
+			"long":   val,
 		})
 	}
 }
